Refactoring: New UsbSerialPort write API
The UsbSerialPort interface now contains two write methods:
* int write(byte[] src, int length, int timeoutMillis) to write as many
bytes as possible within the given timeout and returns the actual number
of bytes written. Is implemented by the drivers.
* void write(final byte[] src) which is implemented in
CommonUsbSerialPort using the other write method. Writes all bytes of
the given array to the port without any timeout and fails if not all
bytes could be transferred.
On this occasion, the driver write implementations were unified and
simplified.